Rename toolbox.ts -> index.ts
